+++ src/main/java/io/github/mzmine/util/spectraldb/parser/MonaJsonParser.java
@@ -1,17 +1,17 @@
 /*
- * Copyright 2006-2018 The MZmine 2 Development Team
+ * Copyright 2006-2020 The MZmine Development Team
  * 
- * This file is part of MZmine 2.
+ * This file is part of MZmine.
  * 
- * MZmine 2 is free software; you can redistribute it and/or modify it under the terms of the GNU
+ * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
  * General Public License as published by the Free Software Foundation; either version 2 of the
  * License, or (at your option) any later version.
  * 
- * MZmine 2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
+ * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
  * 
- * You should have received a copy of the GNU General Public License along with MZmine 2; if not,
+ * You should have received a copy of the GNU General Public License along with MZmine; if not,
  * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
  * USA
  */
@@ -129,7 +129,6 @@ public class MonaJsonParser extends SpectralDBParser {
           i = 0;
         }
 
-
         // to many errors? wrong data format?
         if (parser.size() > 0 && parser.get(0).getError() >= 4) {
           logger.log(Level.WARNING, "This file was no MONA spectral json library");
@@ -145,14 +144,12 @@ public class MonaJsonParser extends SpectralDBParser {
     while (parser.stream().anyMatch(ip -> !ip.isDone)) {
     }
 
-
     if (parser.size() == 1 || parser.get(0).getError() > parser.get(0).getCorrect())
       return false;
     //
     finish();
     return true;
   }
-
 
   /**
    * Test of inner parser to separate reading and parsing
@@ -221,7 +218,6 @@ public class MonaJsonParser extends SpectralDBParser {
     public boolean isDone() {
       return isDone;
     }
-
 
     public SpectralDBEntry getDBEntry(JsonObject main) {
       // extract dps
@@ -349,9 +345,9 @@ public class MonaJsonParser extends SpectralDBParser {
                   // to minutes
                   if (v.endsWith("sec")) {
                     v = v.substring(0, v.length() - 3);
-                    value = Double.parseDouble(v) / 60d;
+                    value = Float.parseFloat(v) / 60f;
                   } else {
-                    value = Double.parseDouble(v);
+                    value = Float.parseFloat(v);
                   }
                 } catch (Exception ex) {
                 }
@@ -435,7 +431,6 @@ public class MonaJsonParser extends SpectralDBParser {
     private String readCompound(JsonObject main, String id) {
       return main.getJsonArray(COMPOUND).getJsonObject(0).getString(id, null);
     }
-
 
     public DataPoint[] getDataPoints(JsonObject main) {
       String spec = main.getString("spectrum");
